Added FLAC decoder with 24bit & seeking (WIP)
this is a C# port of jFlac with some modifications and added a seek
function ported from original C code and heavily modified. Really really
REALLY messy but kinda works and supports 24bit FLAC so it is better
than FlacBox... I don't have time to fix it right now so I hope it works
kinda smoothly.